pyproject.toml
[project]
name = "devtools-sup"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.13"
dependencies = [
    "dnspython>=2.7.0",
    "flask>=3.1.1",
    "flask-cors>=5.0.0",
    "gunicorn>=23.0.0",
    "pandas>=2.3.1",
    "Pillow>=11.0.0",
    "reportlab>=4.2.0",
]

[dependency-groups]
dev = [
    "ruff>=0.12.3",
]
